requirement.txt
streamlit
crewai
langchain
openai
google-cloud-aiplatform
duckduckgo-search
google-search-results
python-dotenv
langchain_community